pyproject.toml
[project]
name = "negotiation-challenge"
version = "0.1.0"
description = "Local test harness for the Optimization Arena negotiation challenge"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "google-genai>=1.0",
    "rich>=13.0",
]

[project.scripts]
negotiate = "negotiate.cli:cli"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/negotiate"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
